+++ ui/pages/bridge/utils/quote.ts
@@ -45,6 +45,62 @@ export const formatCurrencyAmount = (
   return formatCurrency(amount.toString(), currency, precision);
 };
 
+/**
+ * Formats network fees with dynamic precision to avoid showing $0.00 for non-zero fees.
+ * If fees are less than $0.01, increases decimal places up to 4 until the first non-zero digit is shown.
+ * If fees are non-zero but smaller than $0.0001, rounds up to $0.0001.
+ *
+ * @param stringifiedDecAmount - The fee amount as a string
+ * @param currency - The currency code (e.g., 'USD')
+ * @returns Formatted currency string with appropriate precision
+ */
+export function formatNetworkFee(
+  stringifiedDecAmount: string | null | undefined,
+  currency: string,
+): string | undefined {
+  if (!stringifiedDecAmount) {
+    return undefined;
+  }
+
+  const amount = new BigNumber(stringifiedDecAmount);
+
+  // If amount is zero, return formatted zero
+  if (amount.isZero()) {
+    return formatCurrency(amount.toString(), currency, 2);
+  }
+
+  // If amount is >= $0.01, use standard 2 decimal places
+  if (amount.gte(0.01)) {
+    return formatCurrency(amount.toString(), currency, 2);
+  }
+
+  // For amounts < $0.01, find the precision that shows the first non-zero digit
+  // Try precision from 2 to 4 (max allowed)
+  for (let precision = 2; precision <= 4; precision++) {
+    // Scale the amount by 10^precision to move the target digit to the ones place
+    // Example: 0.0005 with precision 3 → 0.5 (moves 3rd decimal to ones place)
+    const scaleFactor = new BigNumber(10).pow(precision);
+    const scaledAmount = amount.times(scaleFactor);
+
+    // Round using ROUND_HALF_UP to match currency formatter's rounding behavior
+    // If the rounded value is non-zero, this precision will show a non-zero digit
+    const roundedValue = scaledAmount.round(0, BigNumber.ROUND_HALF_UP);
+
+    if (roundedValue.gt(0)) {
+      return formatCurrency(amount.toString(), currency, precision);
+    }
+  }
+
+  // If after 4 decimal places it's still showing as zero but original amount > 0,
+  // round up to $0.0001
+  if (amount.gt(0)) {
+    return formatCurrency('0.0001', currency, 4);
+  }
+
+  // Fallback to 2 decimal places
+  return formatCurrency(amount.toString(), currency, 2);
+}
+
 export const formatProviderLabel = (args?: {
   bridgeId: QuoteResponse['quote']['bridgeId'];
   bridges: QuoteResponse['quote']['bridges'];
